fix(gates): split ext-fetch request fields and stdlib fetch body helpers for the file cap
